A lot of code changes. Now it's more elegant
git-svn-id: https://clonekeenplus.svn.sourceforge.net/svnroot/clonekeenplus/cgenius/trunk@781 4df4b0f3-56ce-47cb-b001-ed939b7d65a6

+++ src/common/CObject.cpp
@@ -301,7 +301,7 @@ bool CObject::checkSolidR(stTile *TileProperty, CMap *p_map, int x2, int y1, int
 				return true;
 		}
 	}
-	if( x2 > ((p_map->m_width)<<CSF) ) return true; // Out of map?
+	if( (Uint16)x2 > ((p_map->m_width)<<CSF) ) return true; // Out of map?
 
 	return false;
 }
@@ -349,7 +349,7 @@ bool CObject::checkSolidD(stTile *TileProperty, CMap *p_map, int x1, int x2, int
 				return true;
 		}
 	}
-	if( y2 > ((p_map->m_height-1)<<CSF) ) return true; // Out of map?
+	if( (Uint16)y2 > ((p_map->m_height-1)<<CSF) ) return true; // Out of map?
 
 	return false;
 }
